Only download artifacts which aren't present

+++ dev-tools/mage/downloads/utils.go
@@ -5,6 +5,7 @@
 package downloads
 
 import (
+	"context"
 	"fmt"
 	"io"
 	"net/http"
@@ -23,6 +24,45 @@ type downloadRequest struct {
 // It writes to the destination file as it downloads it, without
 // loading the entire file into memory.
 func downloadFile(downloadRequest *downloadRequest) error {
+	stat, _ := os.Stat(downloadRequest.TargetPath)
+
+	exp := getExponentialBackoff(3)
+
+	retryCount := 1
+	var fileReader io.ReadCloser
+	download := func() error {
+		req, err := http.NewRequestWithContext(context.Background(), http.MethodGet, downloadRequest.URL, nil)
+		if err != nil {
+			return fmt.Errorf("creating request: %w", err)
+		}
+		// if the target file already exists, add the If-Modified-Since header
+		if stat != nil {
+			req.Header.Add("If-Modified-Since", stat.ModTime().Format(http.TimeFormat))
+		}
+
+		resp, err := http.DefaultClient.Do(req) //nolint:bodyclose // we do close this outside of the function
+		if err != nil {
+			retryCount++
+			return fmt.Errorf("downloading file %s: %w", downloadRequest.URL, err)
+		}
+
+		if resp.StatusCode == http.StatusNotModified {
+			return nil
+		}
+
+		fileReader = resp.Body
+		return nil
+	}
+
+	err := backoff.Retry(download, exp)
+	if err != nil {
+		return err
+	}
+
+	if fileReader == nil { // file already exists with the same content
+		return nil
+	}
+
 	targetFile, err := os.Create(downloadRequest.TargetPath)
 	if err != nil {
 		return fmt.Errorf("creating file: %w", err)
@@ -30,30 +70,6 @@ func downloadFile(downloadRequest *downloadRequest) error {
 	defer func() {
 		_ = targetFile.Close()
 	}()
-
-	exp := getExponentialBackoff(3)
-
-	retryCount := 1
-	var fileReader io.ReadCloser
-	download := func() error {
-		req, err := http.NewRequest(http.MethodGet, downloadRequest.URL, nil)
-		if err != nil {
-			return fmt.Errorf("creating request: %w", err)
-		}
-		resp, err := http.DefaultClient.Do(req)
-		if err != nil {
-			retryCount++
-			return fmt.Errorf("downloading file %s: %w", downloadRequest.URL, err)
-		}
-
-		fileReader = resp.Body
-		return nil
-	}
-
-	err = backoff.Retry(download, exp)
-	if err != nil {
-		return err
-	}
 
 	_, err = io.Copy(targetFile, fileReader)
 	if err != nil {
